assignment/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
certifi==2020.12.5
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.1
Django==2.0.3
django-import-export==2.0.1
et-xmlfile==1.1.0
filelock==3.0.12
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==2.5.1
pipenv==2020.11.15
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
sqlparse==0.4.1
tablib==0.14.0
virtualenv==20.2.2
virtualenv-clone==0.5.4
xlrd==2.0.1
xlwt==1.3.0
